#1012 extract common class between IterableMappingMethod and StreamMappingMethod
